requirements.txt
numpy>=1.24
matplotlib>=3.7
Pillow>=10.0
PyYAML>=6.0
tqdm>=4.65
